refactor: Change virtual time type from `float` to `double`

+++ simulation/Clock.h
@@ -3,14 +3,14 @@
 class Clock {
 private:
     // Time in years
-    float currentTime = 0.0f;
-    float timeStep;  // How many years to go forward
+    double currentTime = 0.0f;
+    double timeStep;  // How many years to go forward
 
     // How long to delay each tick by in seconds
     float delay = 0.5f;
 
 public:
-    explicit Clock(const float stepSize) {
+    explicit Clock(const double stepSize) {
         timeStep = stepSize;
     }
